requirements.txt
# Core dependency
Pillow>=10.0.0

# Optional: Redis storage backend
# redis>=4.0.0

# Optional: development / testing
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# mypy>=1.0.0
# ruff>=0.1.0